Rename errorHander to error-handler to comply with the style of the project.
